requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.0
urllib3==2.0.3
